+++ template/src/main.rs
@@ -5,7 +5,7 @@ use uefi::prelude::*;
 
 #[entry]
 fn main(_handle: Handle, mut system_table: SystemTable<Boot>) -> Status {
-    uefi_services::init(&mut system_table).unwrap();
+    uefi::helpers::init(&mut system_table).unwrap();
 
     Status::SUCCESS
 }