+++ uefi-services/build.rs
@@ -1,0 +1,3 @@
+fn main() {
+    println!("cargo:warning=`uefi-services` is deprecated. Functionality was moved to `uefi::helpers::init`.");
+}